+++ packages/base/src/dialogs/ProcessingFormDialog.tsx
@@ -2,7 +2,7 @@ import { IDict, IJupyterGISModel } from '@jupytergis/schema';
 import { Dialog } from '@jupyterlab/apputils';
 import * as React from 'react';
 import { BaseForm, IBaseFormProps } from '../formbuilder/objectform/baseform';
-import { DissolveForm } from '../formbuilder/objectform/dissolveProcessForm';
+import { DissolveForm } from '../formbuilder/objectform/processform/dissolveProcessForm';
 import { Signal } from '@lumino/signaling';
 import { PromiseDelegate } from '@lumino/coreutils';
