+++ packages/base/src/formbuilder/index.ts
@@ -1,6 +1,6 @@
 export * from './objectform/baseform';
-export * from './objectform/geojsonsource';
-export * from './objectform/layerform';
-export * from './objectform/tilesourceform';
-export * from './objectform/vectorlayerform';
+export * from './objectform/sourceform/geojsonsource';
+export * from './objectform/layerform/layerform';
+export * from './objectform/sourceform/tilesourceform';
+export * from './objectform/layerform/vectorlayerform';
 export * from './creationform';
