+++ packages/base/src/formbuilder/objectform/sourceform/sourceform.ts
@@ -1,5 +1,5 @@
 import { IDict, SourceType } from '@jupytergis/schema';
-import { BaseForm, IBaseFormProps } from './baseform';
+import { BaseForm, IBaseFormProps } from '../baseform';
 import { Signal } from '@lumino/signaling';
 import { IChangeEvent } from '@rjsf/core';
